requirements.txt
apiai==1.2.3
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.2
click==6.7
configparser==3.5.0
dialogflow==0.4.0
dill==0.2.8.2
Flask==1.0.2
future==0.16.0
google-api-core==1.2.1
google-auth==1.5.0
googleapis-common-protos==1.5.3
grpcio==1.13.0
gunicorn==19.9.0
idna==2.7
imgurpython==1.1.7
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.7.2
lxml==4.2.3
marisa-trie==0.7.4
MarkupSafe==1.0
mathparse==0.1.1
nltk==3.3
numpy==1.14.5
oauthlib==2.1.0
pandas==0.23.3
protobuf==3.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pymongo==3.7.0
pythainlp==1.6.0.7
python-dateutil==2.6.1
python-twitter==3.4.2
pytz==2018.5
PyYAML==3.12
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.2.9
urllib3==1.23
Werkzeug==0.14.1
